% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_unique.R
\name{sort_unique}
\alias{sort_unique}
\title{Extract and sort unique elements alphabetically}
\usage{
sort_unique(x)
}
\value{
A character vector.
}
\description{
`sort_unique()` returns a character vector with
the unique elements sorted by alphabetical order.
}
