% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_days_per_emp.R
\name{sum_days_per_emp}
\alias{sum_days_per_emp}
\title{Sum the transaction days by employee.}
\usage{
sum_days_per_emp(.df)
}
\arguments{
\item{.data}{A data frame with column `code` containing one unique value.}
}
\value{
A tibble of employee transaction data.
}
\description{
`sum_days_per_emp()` summarizes
the employee transaction data for one set of transaction codes.
}
