% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transApp-package.R
\docType{package}
\name{transApp-package}
\alias{transApp}
\alias{transApp-package}
\title{transApp: Employee Transaction Application}
\description{
An interactive application of employee transaction data from Megamation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asadow/transApp}
  \item Report bugs at \url{https://github.com/asadow/transApp/issues}
}

}
\author{
\strong{Maintainer}: Adam Sadowski \email{asadowsk@uoguelph.ca}

}
\keyword{internal}
